Quebrar código em de participação em classes e serviços com injeção de dependência
