+++ BaseDeProjetos/Helpers/ParticipacaoHelper.cs
@@ -2,6 +2,7 @@
 using BaseDeProjetos.Models;
 using BaseDeProjetos.Models.DTOs;
 using BaseDeProjetos.Models.Enums;
+using BaseDeProjetos.Models.Helpers;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -59,6 +60,21 @@ namespace BaseDeProjetos.Helpers
         public static List<Prospeccao> FiltrarProspeccoesConvertidas(List<Prospeccao> prospeccoes)
         {
             return prospeccoes.Where(p => p.Status.Any(f => f.Status == StatusProspeccao.Convertida)).ToList();
+        }
+
+        /// <summary>
+        /// Filtra as prospecções do usuario de acordo com a data de inicio e data de fim do filtro
+        /// </summary>
+        /// <param name="prospeccoes"></param>
+        /// <param name="dataInicio"></param>
+        /// <param name="dataFim"></param>
+        public static void PeriodizarProspeccoesUsuario(ProspeccoesUsuarioParticipacao prospeccoes, DateTime dataInicio, DateTime dataFim)
+        {
+            prospeccoes.ProspeccoesMembro = ParticipacaoHelper.FiltrarProspeccoesPorPeriodo(dataInicio, dataFim, prospeccoes.ProspeccoesMembro);
+            prospeccoes.ProspeccoesLider = ParticipacaoHelper.FiltrarProspeccoesPorPeriodo(dataInicio, dataFim, prospeccoes.ProspeccoesLider);
+            prospeccoes.ProspeccoesTotais = ParticipacaoHelper.FiltrarProspeccoesPorPeriodo(dataInicio, dataFim, prospeccoes.ProspeccoesTotais);
+            prospeccoes.ProspeccoesTotaisComProposta = ParticipacaoHelper.FiltrarProspeccoesPorPeriodo(dataInicio, dataFim, prospeccoes.ProspeccoesTotaisComProposta);
+            prospeccoes.ProspeccoesTotaisConvertidas = ParticipacaoHelper.FiltrarProspeccoesPorPeriodo(dataInicio, dataFim, prospeccoes.ProspeccoesTotaisConvertidas);
         }
     }
 }
